requirements.txt
pyTelegramBotAPI
openai==0.28.1
python-dotenv
gTTS
